Ensure query string information isn't kept when redirecting through auth login
Prior to this change query information is lost when going through an auth login. For example if the router `beforeEach` is called on a link sent to a user like: https://mysite.com/a_page?foo=bar

to.path => '/a_page'
to.fullPath => /a_page?foo=bar'

So this ensures that query information is passed along in the target and eventually used after successful authentication.